requirements.txt
azure-ai-textanalytics==5.3.0
# azure-cognitiveservices-vision-contentmoderator==1.0.0
azure-common==1.1.28
azure-core==1.30.2
azure-search-documents==11.4.0
# google-auth==2.31.0
# googleapis-common-protos==1.63.2
grpcio==1.64.1
openai==1.35.3
pinecone-client==4.1.2
pinecone-plugin-interface==0.0.7
protoc-gen-openapiv2==0.0.1
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
# uvicorn==0.30.1
